+++ events/messageCreate.js
@@ -6,19 +6,24 @@ module.exports = {
     execute(message) {
         if (message.author.id !== message.client.user.id) {
 
-                // TODO proper error check of if bot has perms to send message.
-                if (message.content.substring(0, 2) === 'ES') {
-                    message.reply('Hey! Sorry for the inconvenience! We\'ve recently converted the bot completely to slash' +
-                        ' commands (Check it out with `/help`!) to comply with Discord\'s upcoming changes.' +
-                        '\nYou can find out more about those changes at ' +
-                        '<https://support-dev.discord.com/hc/en-us/articles/4404772028055-Message-Content-Access-Deprecation-for-Verified-Bots>' +
-                        '\nAlso, this update was quite a big update so if you notice any bugs, please report' +
-                        ' them in our Support Server: https://discord.gg/XaeERFAVfb' +
-                        '\nThe server is also a good way to keep up to date with bot changelogs or make feature requests.' +
-                        '\n\n**YOU HAVE TO RE-AUTH THE BOT TO THE SERVER FOR SLASH COMMAND ACCESS! Please re-auth the bot by clicking:** ' +
-                        '\nhttps://discord.com/api/oauth2/authorize?client_id=757326308547100712&permissions=93248&scope=applications.commands%20bot')
-                        .catch(e => console.error(e.toString(), `for guild id ${message.guildId}. Can\'t reply ES warning on messageCreate. Bot does not have permissions to send message.`))
-                }
+            if (message.content.substring(0, 2) === 'ES') {
+                let esWarning = 'Hey! Sorry for the inconvenience! We\'ve recently converted the bot completely to slash' +
+                    ' commands (Check it out with `/help`!) to comply with Discord\'s upcoming changes.' +
+                    '\nYou can find out more about those changes at ' +
+                    '<https://support-dev.discord.com/hc/en-us/articles/4404772028055-Message-Content-Access-Deprecation-for-Verified-Bots>' +
+                    '\nAlso, this update was quite a big update so if you notice any bugs, please report' +
+                    ' them in our Support Server: https://discord.gg/XaeERFAVfb' +
+                    '\nThe server is also a good way to keep up to date with bot changelogs or make feature requests.' +
+                    '\n\n**IF YOU\'RE THE SERVER OWNER, YOU HAVE TO RE-AUTH THE BOT TO THE SERVER FOR SLASH COMMAND ACCESS! Please re-auth the bot by clicking:** ' +
+                    '\nhttps://discord.com/api/oauth2/authorize?client_id=757326308547100712&permissions=93248&scope=applications.commands%20bot'
+
+                message.reply(esWarning)
+                    .catch(e => {
+                        console.error(e.toString(), `for guild id ${message.guildId}. Can\'t reply ES warning on messageCreate. Bot does not have permissions to send message.`)
+                        message.author.send(esWarning)
+                            .catch(e => console.error(e.toString(), `for guild id ${message.guildId}. Can\'t DM user.`))
+                    })
+            }
 
 
             try {