+++ deploy-commands.js
@@ -16,7 +16,8 @@ const rest = new REST({ version: '9' }).setToken(token);
 (async () => {
     try {
         await rest.put(
-            Routes.applicationGuildCommands(clientId, guildId),            //Routes.applicationCommands(clientId), // <-- Global version
+            //Routes.applicationGuildCommands(clientId, guildId),   // <-- Guild version
+            Routes.applicationCommands(clientId),                   // <-- Global version
             {body: commands},
         );