+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - `displaystats.js` Fixed empty output if no records in databases.
 - `messageCreate.js` Fixed send messages error for old command usage warning
 - `leaderboard.js` Fixed displaying empty leaderboard.
+- `displaystats.js` Fixed bug where displaystats tries to call a deprecated column.
 
 ## v1.4.3 - 2021-09-28